Create README.md
This is a fairly simple arcade-style space shooter in the style of the early '80's but with original game play (specifically ai mechanics) and assets.  The game binaries are available at:
https://jaredbgreat.itch.io/ship-swarm-shoot-up

It is an original and straight-forward game, and is complete at this time.  It is based on my own simple arcade framework, found here:
https://github.com/BlackJar72/SwingArcadeFramework

I am now releasing the source under the GNU General Public License version 3.0: 
https://www.gnu.org/licenses/gpl-3.0-standalone.html

I reserve the right to make any commercial (i.e., non-open source) derivatives for myself as the copyright holder.
Copyright (C) Jared Blackburn, 2017.